+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Disable the experimental feature 'Apply colors to global skin' by default
 - Only show title tab on object's root (If the permissions tab is available)
 - Support Certificate plugin
+- Global config
 
 ## [1.1.0]
 - View: Tile or list